Updated Gamma Prez
Added a few more slides to the gamma slides.  These are actually from my powerbi, which I will also be updating.  The slides focus on claims and drug costs with state slicers to dig deeper.  There is a ton more to look at and will add very soon....I hope